Refactor folders structure and update some file names
