+++ src/components/QRCodeScanner/QRCodeScanner.js
@@ -3,7 +3,7 @@ import Modal from 'react-modal';
 import Webcam from 'react-webcam';
 import { BsQrCodeScan } from 'react-icons/bs';
 import { PiCameraRotateFill } from 'react-icons/pi';
-import Spinner from '../Spinner';
+import Spinner from '../Shared/Spinner';
 import { useTranslation } from 'react-i18next';
 import { FaCheckCircle } from "react-icons/fa";
 import { RiZoomInFill, RiZoomOutFill } from "react-icons/ri";
