+++ src/pages/AddCredentials/AddCredentials.js
@@ -7,10 +7,10 @@ import { useQRScanner } from '../../hooks/useQRScanner';
 import QRCodeScanner from '../../components/QRCodeScanner/QRCodeScanner';
 import RedirectPopup from '../../components/Popups/RedirectPopup';
 import QRButton from '../../components/Buttons/QRButton';
-import { H1 } from '../../components/Heading';
+import { H1 } from '../../components/Shared/Heading';
 import ContainerContext from '../../context/ContainerContext';
 import Button from '../../components/Buttons/Button';
-import PageDescription from '../../components/PageDescription';
+import PageDescription from '../../components/Shared/PageDescription';
 
 function highlightBestSequence(issuer, search) {
 	if (typeof issuer !== 'string' || typeof search !== 'string') {
